Move objreg to api
