+++ vimiv/api/__init__.py
@@ -6,4 +6,4 @@
 
 """TODO"""
 
-from . import status
+from . import objreg, status
