+++ vimiv/api/objreg.py
